pyproject.toml
[project]
name = "nebula-pyg"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "caifengze", email = "fengzcw@gmail.com"},
    {name = "fengze", email = ""},
    {name = "", email = ""},
    {name = "fengzc", email = ""},
]
dependencies = [
    "nebula3-python @ file:///${PROJECT_ROOT}/../nebula-python",
    "numpy>=2.3.2",
    "pytest>=8.4.1",
    "pip>=25.1.1",
    "torchvision>=0.22.1",
    "torchaudio>=2.7.1",
    "torch-geometric>=2.6.1",
    "torch==2.7.1+cpu",
    "pyg-lib @ https://data.pyg.org/whl/torch-2.7.0%2Bcpu/pyg_lib-0.4.0%2Bpt27cpu-cp311-cp311-linux_x86_64.whl",
    "tqdm>=4.67.1",
    "ogb>=1.3.6",
]
requires-python = ">=3.6.2,<3.13"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = true

[[tool.pdm.source]]
name = "private"
url = "https://download.pytorch.org/whl/cpu"
include_packages = ["torch", "torchvision", "torchaudio"]
exclude_packages = ["*"]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[dependency-groups]
dev = [
    "black>=25.1.0",
]
